pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "courier-dtn"
version = "0.1.0"
description = "Delay/Disruption-tolerant reliable file transfer over UDP"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Miftah Meky", email = "miftah.m.meky.27@dartmouth.edu" },
]

dependencies = [
  "pyyaml>=6.0",
]

[project.scripts]
courier = "courier.cli:main"

[tool.setuptools]
packages = ["courier"]